+++ src/nimib.nim
@@ -79,6 +79,11 @@ template nbCodeInBlock*(body: untyped): untyped =
     nbCode:
       body
 
+template nimibCode*(body: untyped) =
+  newNbCodeBlock("nimibCode", body):
+    discard
+  body
+
 template nbText*(text: string) =
   newNbSlimBlock("nbText"):
     nb.blk.output = text
@@ -130,23 +135,10 @@ template nbRawHtml*(content: string) =
     nb.blk.output = content
 
 template nbJsFromStringInit*(body: string): NbBlock =
-  var result = NbBlock(command: "nbCodeToJs", code: body, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
+  var result = NbBlock(command: "nbJsFromCode", code: body, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
   result.context["transformedCode"] = body
+  result.context["putAtTop"] = false
   result
-
-template nbJsFromCodeInit*(args: varargs[untyped]): NbBlock =
-  let (code, originalCode) = nimToJsString(true, args)
-  var result = NbBlock(command: "nbCodeToJs", code: originalCode, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
-  result.context["transformedCode"] = code
-  result
-
-template nbCodeToJsInit*(args: varargs[untyped]): NbBlock {.deprecated: "Use nbJsFromCodeInit or nbJsFromStringInit instead".} =
-  nbJsFromCodeInit(args)
-
-template addCodeToJs*(script: NbBlock, args: varargs[untyped]) =
-  let (code, originalCode) = nimToJsString(false, args)
-  script.code &= "\n" & originalCode
-  script.context["transformedCode"] = script.context["transformedCode"].vString & "\n" & code
 
 template addStringToJs*(script: NbBlock, body: string) =
   script.code &= "\n" & body
@@ -161,8 +153,31 @@ template nbJsFromString*(body: string) =
   script.addToDocAsJs
 
 template nbJsFromCode*(args: varargs[untyped]) =
-  let script = nbJsFromCodeInit(args)
-  script.addToDocAsJs
+  let (code, originalCode) = nimToJsString(putCodeInBlock=false, args)
+  var result = NbBlock(command: "nbJsFromCode", code: originalCode, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
+  result.context["transformedCode"] = code
+  result.context["putAtTop"] = false
+  result.addToDocAsJs
+
+template nbJsFromCodeInBlock*(args: varargs[untyped]) =
+  let (code, originalCode) = nimToJsString(putCodeInBlock=true, args)
+  var result = NbBlock(command: "nbJsFromCode", code: originalCode, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
+  result.context["transformedCode"] = code
+  result.context["putAtTop"] = false
+  result.addToDocAsJs
+
+template nbJsFromCodeGlobal*(args: varargs[untyped]) =
+  let (code, originalCode) = nimToJsString(putCodeInBlock=false, args)
+  var result = NbBlock(command: "nbJsFromCode", code: originalCode, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
+  result.context["transformedCode"] = code
+  result.context["putAtTop"] = true
+  result.addToDocAsJs
+
+template nbJsFromCodeOwnFile*(args: varargs[untyped]) =
+  let (code, originalCode) = nimToJsString(putCodeInBlock=false, args)
+  var result = NbBlock(command: "nbJsFromCodeOwnFile", code: originalCode, context: newContext(searchDirs = @[], partials = nb.partials), output: "")
+  result.context["transformedCode"] = code
+  result.addToDocAsJs
 
 template nbCodeToJs*(args: varargs[untyped]) {.deprecated: "Use nbJsFromCode or nbJsFromString instead".} =
   nbJsFromCode(args)
@@ -171,7 +186,7 @@ template nbCodeToJs*(args: varargs[untyped]) {.deprecated: "Use nbJsFromCode or 
 when moduleAvailable(karax/kbase):
   template nbKaraxCode*(args: varargs[untyped]) =
     let rootId = "karax-" & $nb.newId()
-    nbRawOutput: "<div id=\"" & rootId & "\"></div>"
+    nbRawHtml: "<div id=\"" & rootId & "\"></div>"
     nbKaraxCodeBackend(rootId, args)
 
 template nbJsShowSource*(message: string = "") =
@@ -193,6 +208,8 @@ template nbSave* =
   #   - in memory partial contains default mustache assets
   #   - to override/customize (for a bunch of documents) the best way is to modify a version on file
   #   - in case you need to manage additional exceptions for a specific document add a new set of partials before calling nbSave
+  nb.nbCollectAllNbJs()
+
   nb.context.searchDirs(nb.templateDirs)
   nb.context.searchTable(nb.partials)